renamed config test file
